+++ .copr/getsources.sh
@@ -20,21 +20,25 @@ then
     # So we'll do a little bash magic to get the folder names. Then we'll extract everything, rename as appropriate,
     # and rebuild a full source tarball that includes the submodule.
     cd "$workingdir" || exit
-    curl -L "$CoreRepo/archive/$CoreTag.tar.gz" -o "$CoreTag.tar.gz"
-    curl -L "$LauncherRepo/archive/$LauncherTag.tar.gz" -o "$LauncherTag.tar.gz"
+    echo "Downloading $CoreTag => xlcore-$CoreTag.tar.gz"
+    curl -L "$CoreRepo/archive/$CoreTag.tar.gz" -o "xlcore-$CoreTag.tar.gz"
+    echo "Downloading $LauncherTag => launcher-$LauncherTag.tar.gz"
+    curl -L "$LauncherRepo/archive/$LauncherTag.tar.gz" -o "launcher-$LauncherTag.tar.gz"
     # These next two lines will get the folder names at the top of the tarball
-    CoreDir=$(tar -tf "$CoreTag.tar.gz" | head -n 1)
-    LauncherDir=$(tar -tf "$LauncherTag.tar.gz" | head -n 1)
+    CoreDir=$(tar -tf "xlcore-$CoreTag.tar.gz" | head -n 1)
+    LauncherDir=$(tar -tf "launcher-$LauncherTag.tar.gz" | head -n 1)
     # Extract XIVLauncher.Core tarball and then rename the folder to XIVLauncher.Core
-    tar -xf "$CoreTag.tar.gz"
+    echo "Extracting xlcore-$CoreTag.tar.gz"
+    tar -xf "xlcore-$CoreTag.tar.gz"
     mv "$CoreDir" XIVLauncher.Core
     # Remove an empty directory that will get in our way
     rmdir XIVLauncher.Core/lib/FFXIVQuickLauncher
     # Extract the FFXIVQuickLauncher tarball to XIVLauncher.Core/lib/, then rename to FFXIVQuickLauncher
-    tar -C XIVLauncher.Core/lib -xf "$LauncherTag.tar.gz"
+    echo "Extracting launcher-$LauncherTag.tar.gz and moving files to lib/FFXIVQuickLauncher"
+    tar -C XIVLauncher.Core/lib -xf "launcher-$LauncherTag.tar.gz"
     mv "XIVLauncher.Core/lib/$LauncherDir" "XIVLauncher.Core/lib/FFXIVQuickLauncher"
     # Create a new, complete tarball.
-    echo "Making tarball XIVLauncher.Core-$CoreTag.tar.gz"
+    echo "Writing XIVLauncher.Core-$CoreTag.tar.gz"
     tar -czf "$source0" XIVLauncher.Core
 fi
 cd "$repodir" || exit
@@ -45,7 +49,7 @@ then
     mkdir -p "$workingdir/XIVLauncher4rpm-$DownstreamTag"
     cp CHANGELOG.md cleanupprofile.sh openssl_fix.cnf README.md xivlauncher.sh XIVLauncher.desktop XIVLauncher-custom.desktop COPYING "$workingdir/XIVLauncher4rpm-$DownstreamTag/"
     cd "$workingdir" || exit
-    echo "Making tarball XIVLauncher4rpm-$DownstreamTag.tar.gz"
+    echo "Writing XIVLauncher4rpm-$DownstreamTag.tar.gz"
     tar -czf "$source1" "XIVLauncher4rpm-$DownstreamTag"
 fi
 cp "$repodir/_version" "$xlsource"